refactor: Move axios instance methods to new instanceMethods.ts file
